backend: initialize database on /INITIALIZE get request
